feat: Created a Repositories struct holding all the repositories
