+++ internal/users/service/user_service_test.go
@@ -7,66 +7,11 @@ import (
 	"time"
 
 	"github.com/KnoblauchPilze/user-service/pkg/communication"
-	"github.com/KnoblauchPilze/user-service/pkg/db"
 	"github.com/KnoblauchPilze/user-service/pkg/errors"
 	"github.com/KnoblauchPilze/user-service/pkg/persistence"
-	"github.com/KnoblauchPilze/user-service/pkg/repositories"
 	"github.com/google/uuid"
 	"github.com/stretchr/testify/assert"
 )
-
-type mockUserRepository struct {
-	repositories.UserRepository
-
-	user      persistence.User
-	ids       []uuid.UUID
-	err       error
-	updateErr error
-
-	createCalled   int
-	createdUser    persistence.User
-	getCalled      int
-	getId          uuid.UUID
-	getEmailCalled int
-	getEmail       string
-	listCalled     int
-	updateCalled   int
-	updatedUser    persistence.User
-	deleteCalled   int
-}
-
-type mockApiKeyRepository struct {
-	repositories.ApiKeyRepository
-
-	apiKeyIds []uuid.UUID
-	createErr error
-	getErr    error
-	deleteErr error
-
-	createCalled       int
-	createdApiKey      persistence.ApiKey
-	getForUserCalled   int
-	getForUserTxCalled int
-	userId             uuid.UUID
-	deleteCalled       int
-	deleteTxCalled     int
-	deleteIds          []uuid.UUID
-}
-
-type mockConnectionPool struct {
-	db.ConnectionPool
-
-	tx  mockTransaction
-	err error
-}
-
-type mockTransaction struct {
-	db.Transaction
-
-	timeStamp time.Time
-
-	closeCalled int
-}
 
 var errDefault = fmt.Errorf("some error")
 var defaultUserId = uuid.MustParse("08ce96a3-3430-48a8-a3b2-b1c987a207ca")
@@ -93,7 +38,7 @@ func TestUserService_Create_CallsRepositoryCreate(t *testing.T) {
 	mur := &mockUserRepository{}
 	mkr := &mockApiKeyRepository{}
 	mc := &mockConnectionPool{}
-	s := NewUserService(Config{}, mc, mur, mkr)
+	s := NewUserService(Config{}, mc, createRepositories(mkr, mur))
 
 	s.Create(context.Background(), defaultUserDtoRequest)
 
@@ -108,7 +53,7 @@ func TestUserService_Create_WhenRepositoryFails_ExpectError(t *testing.T) {
 	}
 	mkr := &mockApiKeyRepository{}
 	mc := &mockConnectionPool{}
-	s := NewUserService(Config{}, mc, mur, mkr)
+	s := NewUserService(Config{}, mc, createRepositories(mkr, mur))
 
 	_, err := s.Create(context.Background(), defaultUserDtoRequest)
 
@@ -123,7 +68,7 @@ func TestUserService_Create_ReturnsCreatedUser(t *testing.T) {
 	}
 	mkr := &mockApiKeyRepository{}
 	mc := &mockConnectionPool{}
-	s := NewUserService(Config{}, mc, mur, mkr)
+	s := NewUserService(Config{}, mc, createRepositories(mkr, mur))
 
 	actual, err := s.Create(context.Background(), defaultUserDtoRequest)
 
@@ -145,7 +90,7 @@ func TestUserService_Get_CallsRepositoryGet(t *testing.T) {
 	mur := &mockUserRepository{}
 	mkr := &mockApiKeyRepository{}
 	mc := &mockConnectionPool{}
-	s := NewUserService(Config{}, mc, mur, mkr)
+	s := NewUserService(Config{}, mc, createRepositories(mkr, mur))
 
 	s.Get(context.Background(), defaultUserId)
 
@@ -160,7 +105,7 @@ func TestUserService_Get_WhenRepositoryFails_ExpectError(t *testing.T) {
 	}
 	mkr := &mockApiKeyRepository{}
 	mc := &mockConnectionPool{}
-	s := NewUserService(Config{}, mc, mur, mkr)
+	s := NewUserService(Config{}, mc, createRepositories(mkr, mur))
 
 	_, err := s.Get(context.Background(), defaultUserId)
 
@@ -175,7 +120,7 @@ func TestUserService_Get_ReturnsUser(t *testing.T) {
 	}
 	mkr := &mockApiKeyRepository{}
 	mc := &mockConnectionPool{}
-	s := NewUserService(Config{}, mc, mur, mkr)
+	s := NewUserService(Config{}, mc, createRepositories(mkr, mur))
 
 	actual, err := s.Get(context.Background(), defaultUserId)
 
@@ -198,7 +143,7 @@ func TestUserService_List_CallsRepositoryList(t *testing.T) {
 	mur := &mockUserRepository{}
 	mkr := &mockApiKeyRepository{}
 	mc := &mockConnectionPool{}
-	s := NewUserService(Config{}, mc, mur, mkr)
+	s := NewUserService(Config{}, mc, createRepositories(mkr, mur))
 
 	s.List(context.Background())
 
@@ -213,7 +158,7 @@ func TestUserService_List_WhenRepositoryFails_ExpectError(t *testing.T) {
 	}
 	mkr := &mockApiKeyRepository{}
 	mc := &mockConnectionPool{}
-	s := NewUserService(Config{}, mc, mur, mkr)
+	s := NewUserService(Config{}, mc, createRepositories(mkr, mur))
 
 	_, err := s.List(context.Background())
 
@@ -231,7 +176,7 @@ func TestUserService_List_ReturnsAllUsers(t *testing.T) {
 	}
 	mkr := &mockApiKeyRepository{}
 	mc := &mockConnectionPool{}
-	s := NewUserService(Config{}, mc, mur, mkr)
+	s := NewUserService(Config{}, mc, createRepositories(mkr, mur))
 
 	actual, err := s.List(context.Background())
 
@@ -245,7 +190,7 @@ func TestUserService_Update_CallsRepositoryGetAndUpdate(t *testing.T) {
 	mur := &mockUserRepository{}
 	mkr := &mockApiKeyRepository{}
 	mc := &mockConnectionPool{}
-	s := NewUserService(Config{}, mc, mur, mkr)
+	s := NewUserService(Config{}, mc, createRepositories(mkr, mur))
 
 	s.Update(context.Background(), defaultUserId, defaultUserDtoRequest)
 
@@ -262,7 +207,7 @@ func TestUserService_Update_WhenGetFails_ExpectError(t *testing.T) {
 	}
 	mkr := &mockApiKeyRepository{}
 	mc := &mockConnectionPool{}
-	s := NewUserService(Config{}, mc, mur, mkr)
+	s := NewUserService(Config{}, mc, createRepositories(mkr, mur))
 
 	_, err := s.Update(context.Background(), defaultUserId, defaultUserDtoRequest)
 
@@ -277,7 +222,7 @@ func TestUserService_Update_CallsUpdateWithUpdatedValues(t *testing.T) {
 	}
 	mkr := &mockApiKeyRepository{}
 	mc := &mockConnectionPool{}
-	s := NewUserService(Config{}, mc, mur, mkr)
+	s := NewUserService(Config{}, mc, createRepositories(mkr, mur))
 
 	s.Update(context.Background(), defaultUserId, defaultUserDtoRequest)
 
@@ -299,7 +244,7 @@ func TestUserService_Update_WhenUpdateFails_ExpectError(t *testing.T) {
 	}
 	mkr := &mockApiKeyRepository{}
 	mc := &mockConnectionPool{}
-	s := NewUserService(Config{}, mc, mur, mkr)
+	s := NewUserService(Config{}, mc, createRepositories(mkr, mur))
 
 	_, err := s.Update(context.Background(), defaultUserId, defaultUserDtoRequest)
 
@@ -314,7 +259,7 @@ func TestUserService_Update_ReturnsUpdatedUser(t *testing.T) {
 	}
 	mkr := &mockApiKeyRepository{}
 	mc := &mockConnectionPool{}
-	s := NewUserService(Config{}, mc, mur, mkr)
+	s := NewUserService(Config{}, mc, createRepositories(mkr, mur))
 
 	actual, err := s.Update(context.Background(), defaultUserId, defaultUserDtoRequest)
 
@@ -337,7 +282,7 @@ func TestUserService_Delete_CallsRepositoryDelete(t *testing.T) {
 	mur := &mockUserRepository{}
 	mkr := &mockApiKeyRepository{}
 	mc := &mockConnectionPool{}
-	s := NewUserService(Config{}, mc, mur, mkr)
+	s := NewUserService(Config{}, mc, createRepositories(mkr, mur))
 
 	s.Delete(context.Background(), defaultUserId)
 
@@ -351,7 +296,7 @@ func TestUserService_Delete_CallsTransactionClose(t *testing.T) {
 	mur := &mockUserRepository{}
 	mkr := &mockApiKeyRepository{}
 	mc := &mockConnectionPool{}
-	s := NewUserService(Config{}, mc, mur, mkr)
+	s := NewUserService(Config{}, mc, createRepositories(mkr, mur))
 
 	s.Delete(context.Background(), defaultUserId)
 
@@ -366,7 +311,7 @@ func TestUserService_Delete_WhenCreatingTransactionFails_ExpectError(t *testing.
 	mc := &mockConnectionPool{
 		err: errDefault,
 	}
-	s := NewUserService(Config{}, mc, mur, mkr)
+	s := NewUserService(Config{}, mc, createRepositories(mkr, mur))
 
 	err := s.Delete(context.Background(), defaultUserId)
 
@@ -379,7 +324,7 @@ func TestUserService_Delete_FetchesUsersKeys(t *testing.T) {
 	mur := &mockUserRepository{}
 	mkr := &mockApiKeyRepository{}
 	mc := &mockConnectionPool{}
-	s := NewUserService(Config{}, mc, mur, mkr)
+	s := NewUserService(Config{}, mc, createRepositories(mkr, mur))
 
 	s.Delete(context.Background(), defaultUserId)
 
@@ -395,7 +340,7 @@ func TestUserService_Delete_DeletesTheRightKeys(t *testing.T) {
 		apiKeyIds: []uuid.UUID{defaultApiKeyId},
 	}
 	mc := &mockConnectionPool{}
-	s := NewUserService(Config{}, mc, mur, mkr)
+	s := NewUserService(Config{}, mc, createRepositories(mkr, mur))
 
 	s.Delete(context.Background(), defaultUserId)
 
@@ -410,7 +355,7 @@ func TestUserService_Delete_WhenUserRepositoryFails_ExpectError(t *testing.T) {
 	}
 	mkr := &mockApiKeyRepository{}
 	mc := &mockConnectionPool{}
-	s := NewUserService(Config{}, mc, mur, mkr)
+	s := NewUserService(Config{}, mc, createRepositories(mkr, mur))
 
 	err := s.Delete(context.Background(), defaultUserId)
 
@@ -425,7 +370,7 @@ func TestUserService_Delete_WhenApiKeyRepositoryGetFails_ExpectError(t *testing.
 		getErr: errDefault,
 	}
 	mc := &mockConnectionPool{}
-	s := NewUserService(Config{}, mc, mur, mkr)
+	s := NewUserService(Config{}, mc, createRepositories(mkr, mur))
 
 	err := s.Delete(context.Background(), defaultUserId)
 
@@ -440,7 +385,7 @@ func TestUserService_Delete_WhenApiKeyRepositoryDeleteFails_ExpectError(t *testi
 		deleteErr: errDefault,
 	}
 	mc := &mockConnectionPool{}
-	s := NewUserService(Config{}, mc, mur, mkr)
+	s := NewUserService(Config{}, mc, createRepositories(mkr, mur))
 
 	err := s.Delete(context.Background(), defaultUserId)
 
@@ -453,7 +398,7 @@ func TestUserService_Delete_WhenRepositoriesSucceeds_ExpectSuccess(t *testing.T)
 	mur := &mockUserRepository{}
 	mkr := &mockApiKeyRepository{}
 	mc := &mockConnectionPool{}
-	s := NewUserService(Config{}, mc, mur, mkr)
+	s := NewUserService(Config{}, mc, createRepositories(mkr, mur))
 
 	err := s.Delete(context.Background(), defaultUserId)
 
@@ -466,7 +411,7 @@ func TestUserService_Login_FetchesUserByEmail(t *testing.T) {
 	mur := &mockUserRepository{}
 	mkr := &mockApiKeyRepository{}
 	mc := &mockConnectionPool{}
-	s := NewUserService(Config{}, mc, mur, mkr)
+	s := NewUserService(Config{}, mc, createRepositories(mkr, mur))
 
 	s.Login(context.Background(), defaultUserDtoRequest)
 
@@ -482,7 +427,7 @@ func TestUserService_Login_WhenGetUserFails_ExpectError(t *testing.T) {
 	}
 	mkr := &mockApiKeyRepository{}
 	mc := &mockConnectionPool{}
-	s := NewUserService(Config{}, mc, mur, mkr)
+	s := NewUserService(Config{}, mc, createRepositories(mkr, mur))
 
 	_, err := s.Login(context.Background(), defaultUserDtoRequest)
 
@@ -497,7 +442,7 @@ func TestUserService_Login_WhenCredentialsAreWrong_ExpectError(t *testing.T) {
 	}
 	mkr := &mockApiKeyRepository{}
 	mc := &mockConnectionPool{}
-	s := NewUserService(Config{}, mc, mur, mkr)
+	s := NewUserService(Config{}, mc, createRepositories(mkr, mur))
 
 	userRequest := communication.UserDtoRequest{
 		Email:    defaultUserEmail,
@@ -520,7 +465,7 @@ func TestUserService_Login_CreatesApiKeyForUser(t *testing.T) {
 	c := Config{
 		ApiKeyValidity: 1 * time.Hour,
 	}
-	s := NewUserService(c, mc, mur, mkr)
+	s := NewUserService(c, mc, createRepositories(mkr, mur))
 
 	s.Login(context.Background(), defaultUserDtoRequest)
 
@@ -540,7 +485,7 @@ func TestUserService_Login_WhenApiKeyCreationFails_ExpectError(t *testing.T) {
 		createErr: errDefault,
 	}
 	mc := &mockConnectionPool{}
-	s := NewUserService(Config{}, mc, mur, mkr)
+	s := NewUserService(Config{}, mc, createRepositories(mkr, mur))
 
 	_, err := s.Login(context.Background(), defaultUserDtoRequest)
 
@@ -559,7 +504,7 @@ func TestUserService_Login_ReturnsCreatedKey(t *testing.T) {
 	c := Config{
 		ApiKeyValidity: 1 * time.Hour,
 	}
-	s := NewUserService(c, mc, mur, mkr)
+	s := NewUserService(c, mc, createRepositories(mkr, mur))
 
 	actual, err := s.Login(context.Background(), defaultUserDtoRequest)
 
@@ -574,7 +519,7 @@ func TestUserService_LoginById_FetchesUser(t *testing.T) {
 	mur := &mockUserRepository{}
 	mkr := &mockApiKeyRepository{}
 	mc := &mockConnectionPool{}
-	s := NewUserService(Config{}, mc, mur, mkr)
+	s := NewUserService(Config{}, mc, createRepositories(mkr, mur))
 
 	s.LoginById(context.Background(), defaultUserId)
 
@@ -590,7 +535,7 @@ func TestUserService_LoginById_WhenGetUserFails_ExpectError(t *testing.T) {
 	}
 	mkr := &mockApiKeyRepository{}
 	mc := &mockConnectionPool{}
-	s := NewUserService(Config{}, mc, mur, mkr)
+	s := NewUserService(Config{}, mc, createRepositories(mkr, mur))
 
 	_, err := s.LoginById(context.Background(), defaultUserId)
 
@@ -608,7 +553,7 @@ func TestUserService_LoginById_CreatesApiKeyForUser(t *testing.T) {
 	c := Config{
 		ApiKeyValidity: 1 * time.Hour,
 	}
-	s := NewUserService(c, mc, mur, mkr)
+	s := NewUserService(c, mc, createRepositories(mkr, mur))
 
 	s.LoginById(context.Background(), defaultUserId)
 
@@ -626,7 +571,7 @@ func TestUserService_LoginById_WhenApiKeyCreationFails_ExpectError(t *testing.T)
 		createErr: errDefault,
 	}
 	mc := &mockConnectionPool{}
-	s := NewUserService(Config{}, mc, mur, mkr)
+	s := NewUserService(Config{}, mc, createRepositories(mkr, mur))
 
 	_, err := s.LoginById(context.Background(), defaultUserId)
 
@@ -645,7 +590,7 @@ func TestUserService_LoginById_ReturnsCreatedKey(t *testing.T) {
 	c := Config{
 		ApiKeyValidity: 1 * time.Hour,
 	}
-	s := NewUserService(c, mc, mur, mkr)
+	s := NewUserService(c, mc, createRepositories(mkr, mur))
 
 	actual, err := s.LoginById(context.Background(), defaultUserId)
 
@@ -660,7 +605,7 @@ func TestUserService_Logout_FetchesGetUser(t *testing.T) {
 	mur := &mockUserRepository{}
 	mkr := &mockApiKeyRepository{}
 	mc := &mockConnectionPool{}
-	s := NewUserService(Config{}, mc, mur, mkr)
+	s := NewUserService(Config{}, mc, createRepositories(mkr, mur))
 
 	s.Logout(context.Background(), defaultUserId)
 
@@ -676,7 +621,7 @@ func TestUserService_Logout_WhenGetUserFails_ExpectError(t *testing.T) {
 	}
 	mkr := &mockApiKeyRepository{}
 	mc := &mockConnectionPool{}
-	s := NewUserService(Config{}, mc, mur, mkr)
+	s := NewUserService(Config{}, mc, createRepositories(mkr, mur))
 
 	err := s.Logout(context.Background(), defaultUserId)
 
@@ -689,7 +634,7 @@ func TestUserService_Logout_FetchesUserKeys(t *testing.T) {
 	mur := &mockUserRepository{}
 	mkr := &mockApiKeyRepository{}
 	mc := &mockConnectionPool{}
-	s := NewUserService(Config{}, mc, mur, mkr)
+	s := NewUserService(Config{}, mc, createRepositories(mkr, mur))
 
 	s.Logout(context.Background(), defaultUserId)
 
@@ -705,7 +650,7 @@ func TestUserService_Logout_WhenGetForUserFails_ExpectError(t *testing.T) {
 		getErr: errDefault,
 	}
 	mc := &mockConnectionPool{}
-	s := NewUserService(Config{}, mc, mur, mkr)
+	s := NewUserService(Config{}, mc, createRepositories(mkr, mur))
 
 	err := s.Logout(context.Background(), defaultUserId)
 
@@ -720,7 +665,7 @@ func TestUserService_Logout_DeletesUserKeys(t *testing.T) {
 		apiKeyIds: []uuid.UUID{defaultApiKeyId},
 	}
 	mc := &mockConnectionPool{}
-	s := NewUserService(Config{}, mc, mur, mkr)
+	s := NewUserService(Config{}, mc, createRepositories(mkr, mur))
 
 	err := s.Logout(context.Background(), defaultUserId)
 
@@ -738,7 +683,7 @@ func TestUserService_Logout_WhenDeleteFails_ExpectError(t *testing.T) {
 		deleteErr: errDefault,
 	}
 	mc := &mockConnectionPool{}
-	s := NewUserService(Config{}, mc, mur, mkr)
+	s := NewUserService(Config{}, mc, createRepositories(mkr, mur))
 
 	err := s.Logout(context.Background(), defaultUserId)
 
@@ -752,86 +697,10 @@ func TestUserService_Logout_WhenAlreadyLoggedOut_DoesNotDeleteKeys(t *testing.T)
 	mur := &mockUserRepository{}
 	mkr := &mockApiKeyRepository{}
 	mc := &mockConnectionPool{}
-	s := NewUserService(Config{}, mc, mur, mkr)
+	s := NewUserService(Config{}, mc, createRepositories(mkr, mur))
 
 	err := s.Logout(context.Background(), defaultUserId)
 
 	assert.Equal(0, mkr.deleteCalled)
 	assert.Nil(err)
-}
-
-func (m *mockUserRepository) Create(ctx context.Context, user persistence.User) (persistence.User, error) {
-	m.createCalled++
-	m.createdUser = user
-	return m.user, m.err
-}
-
-func (m *mockUserRepository) Get(ctx context.Context, id uuid.UUID) (persistence.User, error) {
-	m.getCalled++
-	m.getId = id
-	return m.user, m.err
-}
-
-func (m *mockUserRepository) GetByEmail(ctx context.Context, email string) (persistence.User, error) {
-	m.getEmailCalled++
-	m.getEmail = email
-	return m.user, m.err
-}
-
-func (m *mockUserRepository) List(ctx context.Context) ([]uuid.UUID, error) {
-	m.listCalled++
-	return m.ids, m.err
-}
-
-func (m *mockUserRepository) Update(ctx context.Context, user persistence.User) (persistence.User, error) {
-	m.updateCalled++
-	m.updatedUser = user
-	return m.updatedUser, m.updateErr
-}
-
-func (m *mockUserRepository) Delete(ctx context.Context, tx db.Transaction, id uuid.UUID) error {
-	m.deleteCalled++
-	return m.err
-}
-
-func (m *mockApiKeyRepository) Create(ctx context.Context, apiKey persistence.ApiKey) (persistence.ApiKey, error) {
-	m.createCalled++
-	m.createdApiKey = apiKey
-	return apiKey, m.createErr
-}
-
-func (m *mockApiKeyRepository) GetForUser(ctx context.Context, user uuid.UUID) ([]uuid.UUID, error) {
-	m.getForUserCalled++
-	m.userId = user
-	return m.apiKeyIds, m.getErr
-}
-
-func (m *mockApiKeyRepository) GetForUserTx(ctx context.Context, tx db.Transaction, user uuid.UUID) ([]uuid.UUID, error) {
-	m.getForUserTxCalled++
-	m.userId = user
-	return m.apiKeyIds, m.getErr
-}
-
-func (m *mockApiKeyRepository) Delete(ctx context.Context, ids []uuid.UUID) error {
-	m.deleteCalled++
-	m.deleteIds = ids
-	return m.deleteErr
-}
-
-func (m *mockApiKeyRepository) DeleteTx(ctx context.Context, tx db.Transaction, ids []uuid.UUID) error {
-	m.deleteTxCalled++
-	m.deleteIds = ids
-	return m.deleteErr
-}
-
-func (m *mockConnectionPool) StartTransaction(ctx context.Context) (db.Transaction, error) {
-	return &m.tx, m.err
-}
-
-func (m *mockTransaction) Close(ctx context.Context) {
-	m.closeCalled++
-}
-
-func (m *mockTransaction) TimeStamp() time.Time {
-	return m.timeStamp
 }
