+++ pkg/repositories/repositories.go
@@ -1,0 +1,8 @@
+package repositories
+
+type Repositories struct {
+	Acl       AclRepository
+	ApiKey    ApiKeyRepository
+	UserLimit UserLimitRepository
+	User      UserRepository
+}
